Introduce inverse of control for application layer
