+++ api/src/server/rest/controllers/character.ctrl.ts
@@ -8,6 +8,7 @@ import {
 } from '@/application/account-access.service';
 import { Character } from '@/domain/game/character/character';
 import { AvailableRace } from '@/domain/game/character/race';
+import { RepositoryFactory } from '@/server/repository-factory';
 
 interface CreateNewCharacterReq {
   body: {
@@ -30,7 +31,12 @@ export const createNewCharacter: AppRequestHandler<CreateNewCharacterReq> = asyn
   const accountId = req.accountId;
 
   try {
-    const character = await createCharacter(accountId, name, race);
+    const character = await createCharacter(
+      accountId,
+      name,
+      race,
+      RepositoryFactory.Default
+    );
     logger.info(`Character "${name} with race ${race}" created`);
 
     return res.status(status.CREATED).send(mapCharacterToResponse(character));
@@ -43,7 +49,10 @@ export const findCharactersForUsers: AppRequestHandler = async (req, res) => {
   const accountId = req.accountId;
 
   try {
-    const characters = await getCharacters(accountId);
+    const characters = await getCharacters(
+      accountId,
+      RepositoryFactory.Default
+    );
 
     return res.status(status.OK).send(characters.map(mapCharacterToResponse));
   } catch (err) {
